requirements.txt
django
djangorestframework
Pillow
psycopg2-binary
drf-yasg
django-filter
dj-rest-auth